requirements.txt
fastapi
uvicorn
pydantic
prometheus_client
pypdf
pdfplumber
langchain-text-splitters
groq
google-genai
python-dotenv
requests
rank_bm25
qdrant-client
faiss-cpu
numpy
torch
transformers
redis
aiofiles
pytest
